Time Correction for the Experimental CW EPR Time Series.

Providing more accurate time for EPR spectral line/spectrum appearance. It assumes that the middle \(B\) (or \(g\), \(\nu_{\text{MHz}}\)...etc.) of the EPR spectrum is set as the CF (central field) for the spectrum sweep.

Usage

correct_time_Exp_Specs(time.s, Nscans, sweep.time.s)

Arguments

time.s

Numeric value/vector/column in data frame, corresponding to time (in s) at which the individual EPR spectra were recorded (supplied by the EPR acquisition software).

Nscans

Numeric, number of accumulations (number of scans usually denoted as AVGS) for each spectrum in EPR time series.

sweep.time.s

Numeric, time (in s) for recording individual EPR spectrum \(\equiv\) one "accumulation".

Value

Numeric value/vector, corresponding to time at which the middle (\(x\)-axis) of EPR spectrum/spectra were recorded during the kinetic measurements (e.g. radical formation, stability, electrochemical and/or photochemical measurements).

Details

The actual time at the middle/crossing point is different from that recorder by the EPR acquisition software, see below. This is especially important in determining the kinetics of radical generation or decay. Time is recorded according to the following scheme, where "^v" in the scheme denotes the derivative form of an EPR spectrum:

EPR Spectr.				EPR Spectr.

t[1]	——^v——>	t[2]-delay	—->	t[2]	——^v——-> ...

N_scans*swt		delay		N_scans*swt

The recorded times are: t[1],t[2],t[3],... and the N_scans corresponds to number of scans, swt to sweep time for the individual scan. These parameters can be obtained by the readEPR_params_slct_kin or other functions which can read the instrumental parameter files.

See also

Examples

## 12 s recorded by spectrometer, 6 accumulations by the sweep time of 6 s
correct_time_Exp_Specs(12,Nscans = 6,6)
#> [1] 45