Find Intensity Extremes within the EPR/ENDOR Spectrum

Finding the x positions like B (in mT or G)) or g (unitless) or RF (in MHz) of intensity minimum or maximum within the selected region of an EPR/ENDOR spectrum.

Usage

eval_extremeX_Spec(
  data.spectr,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  xlim = NULL,
  extreme = "max"
)

Arguments

data.spectr

EPR/ENDOR spectrum data frame object with magnetic flux density B (in mT or G) or g-Value or RF (in MHz) column/variable and that of the derivative dIepr_over_dB Intensity. Index column may be included as well.

x

Character string pointing to name of the x-axis/column/variable (in the original data.spectr) like magnetic flux density \(B\), \(g\)-Value or \(RF\) (radio frequency), default: x = "B_mT".

Intensity

Character string pointing to name of the intensity column/variable, in the original data.spectr, if other than dIepr_over_dB name/label is used (e.g. for simulated spectra). Default: Intesity = "dIepr_over_dB".

xlim

Numeric vector corresponding to lower and upper limit of the selected x region, such as xlim = c(3495.4,3595.4) (B in G) or xlim = c(12.5,21.2) (RF in MHz) or xlim = c(2.004,2.001) (g dimensionless). Default: xlim = NULL (corresponding to the entire x range).

extreme

Character string with only a two values allowed: "min" or "max" (default).

Value

Numeric value of x-axis variable like B,g,RF corresponding to minimal or maximal intensity within the EPR/ENDOR spectrum or its integrated form.

See also

Examples

## loading TMPD built-in example file:
tmpd.data.file.path <-
  load_data_example(file = "TMPD_specelchem_accu_b.asc")
## reading data:
tmpd.data.file <-
  readEPR_Exp_Specs(path_to_ASC = tmpd.data.file.path,
                    col.names = c("B_G",
                                  "dIepr_over_dB"),
                    qValue = 3500,
                    norm.vec.add = 20,
                    origin = "winepr")
#
## finding maximum and minimum `B` within the entire
## spectral (`B`) range:
eval_extremeX_Spec(data.spectr = tmpd.data.file)
#> [1] 350.04
#
eval_extremeX_Spec(data.spectr = tmpd.data.file,
                   extreme = "min")
#> [1] 350.09
#
## both values can be checked by the interactive
## spectrum:
plot_EPR_Specs2D_interact(tmpd.data.file)