Basic Calculation of \(g\)-Factor

Calculation of \(g\)-factor according to fundamental formula (see Value). The magnetic flux density (B.val) and microwave frequency (nu.val,\(\nu\)) can be defined, having the common units like G (Gauss) mT (millitesla) or T (tesla) as well as GHz or Hz.

Usage

eval_gFactor(nu.val, nu.unit = "GHz", B.val, B.unit = "mT")

Arguments

nu.val

Numeric, microwave Frequency value.

nu.unit

Character string, frequency unit defined by GHz or Hz, default: nu.unit = "GHz".

B.val

Numeric, magnetic flux density value.

B.unit

Character string, magnetic flux density unit in G or mT or T, default: B.unit = "mT".

Value

\(g\)-Value from \((\nu h)/(\mu_{\text{B}} B)\). Physical constants (\(\mu_{\text{B}}\) and \(h\)) are taken from constants package by the syms.

See also

Examples

eval_gFactor(9.8020458,
             nu.unit = "GHz",
             350.214,
             B.unit = "mT")
#> [1] 1.99973
#
eval_gFactor(nu.val = 9.8020458e+9,
             nu.unit = "Hz",
             B.val = 3502.14,
             B.unit = "G")
#> [1] 1.99973
#
eval_gFactor(9.5421,"GHz",0.333251,"T")
#> [1] 2.04579