Absolute Quantification of Radicals/Spins

Estimate the number or concentration of "spins"/radicals/paramagnetic species using the absolute quantitative method by sigmoid integral as well as by instrumental parameters without the need for a standard sample with known concentration/amount of radicals/"spins". The calculation assumes that the sample height middle point, within an EPR tube, matches the cavity/resonator center.

Usage

quantify_EPR_Abs(
  integ.sigmoid.max,
  instrum.params = NULL,
  path_to_dsc_par,
  origin = "xenon",
  qValue = NULL,
  tube.sample.id.mm,
  point.sample.factor = 8.51e-09,
  fill.sample.h.mm,
  eff.cavity.h.mm = 23,
  fn.B1.Bm.fit.y = c(1.00179, -0.00307086, -0.0265409, 0.000297603, 0.000223277,
    -4.53833e-06, -4.1451e-07, 1.89417e-08, -1.48241e-09),
  fn.B1.Bm.fit.y.max = 0.25,
  Norm.const = NULL,
  Temp.K = NULL,
  S = 0.5
)

Arguments

integ.sigmoid.max

Numeric value or vector of the entire EPR spectrum sigmoid integral.

instrum.params

Named numeric vector containing instrumental parameters required for the quantification =>

PmW	power of the MW source in mW

BmmT	modulation amplitude (magnetic flux density modulation, \(B_{\text{m}}\)) in mT

TK	temperature in K

mwGHz	applied microwave frequency in GHz to record the continuous wave (CW) EPR spectrum

Default: instrum.params = NULL because they are primarily extracted from the path_to_dsc_par based on the origin.

path_to_dsc_par

Character string, path (can be also acquired by the file.path) to .DSC/.dsc or .par (depending on the OS, see origin argument) text files including all instrumental parameters from the EPR machine. Default: path_to_dsc_par = NULL in case if the instrum.params is already defined. IF the instrum.params = NULL then BOTH the path_to_dsc_par AS WELL AS the origin MUST BE DEFINED !

origin

Character string, corresponding to software which was used to obtain the EPR spectra on spectrometers, because the files are slightly different, whether they were recorded by the "WinEpr" (origin = "winepr") or by "Xenon". Default: origin = NULL in case no file is used to extract the parameters (i.e. exactly if path_to_dsc_par = NULL).

qValue

Numeric value of the sensitivity Q factor. For the processed EPR spectra by the {eprscope} package the integ.sigmoid.max is usually normalized by the Q value. Therefore, default: qValue = NULL.

tube.sample.id.mm

Numeric value, which equals to internal diameter (in mm) of the tube/cell used for the quantitative EPR experiment.

point.sample.factor

Numeric value corresponding to point sample correction factor, provided by the cavity/probehead manufacturer. Value for the standard Bruker rectangular cavity is set as default.

fill.sample.h.mm

Numeric value, referring to sample height (in mm) within the tube/cell.

eff.cavity.h.mm

Numeric value, which equals to effective cavity/probehead height/length, usually provided by the probehead manufacturer.

fn.B1.Bm.fit.y

Numeric vector (coefficients) of the polynomial degree from 5 to 12. Coefficients for the standard Bruker rectangular cavity are set as default.

fn.B1.Bm.fit.y.max

Numeric value corresponding to maximum value of the polynomial fit. Value for the standard Bruker rectangular cavity is set as default.

Norm.const

Numeric value corresponding to normalization constant (see quantify_EPR_Norm_const). Default: Norm.const = NULL in case if the EPR spectrum was normalized by such constant either upon measurement or during processing. Otherwise it must be provided by the quantify_EPR_Norm_const.

Temp.K

Numeric value, temperature value in K. Because the instrum.params also contains temperature input one may choose which definition (Temp.K or TK) is taken for the calculation. Either Temp.K or TK CAN BE ALSO NULL but NOT BOTH !! In the latter case, default value 298 K is considered.

S

Numeric value, total spin sample quantum number. For radicals S = 0.5 (default).

Value

List of the following quantities:

N.cm

Number of spins per effective centimeter. It is defined as the cm around the maximum, \(\pm 5\,\text{mm}\), of the intensity distribution curve/polynomial fit within the cavity \(f(B_1,B_{\text{m}})\) from the equation above shown in details.

N.cm3

Number of spins per \(\text{cm}^3\).

c.M

Concentration of spins/radicals in \(\text{mol}\,\text{dm}^{-3}\).

Details

There are two approaches how to quantify the number of paramagnetic species/radicals/spins. The relative one needs a standard sample with a known spin number and can be evaluated by the sigmoid integral ratio of the sample under study and that of the standard. While the absolute method do not need the reference sample, it requires a precise cavity signal calibration as well as standardized cell geometry. Both are provided by the EPR instrument and lab-glass manufacturers (see e.g. hirschCapill2023eprscope). In case of absolute quantitative EPR analysis the sigmoid integral (its maximum value), \(I_{\text{sigmoid}}\),can be used to calculate the number of "spins"/radicals/paramagnetic species, \(N_{\text{Spins}}\) eatonQepr2010,RWeberXenon2011eprscope => $$N_{\text{Spins}} = I_{\text{sigmoid}}\,/\,((c/f(B_1,B_{\text{m}}))\,(G_{\text{R}}\,t_{\text{C}} \,N_{\text{Scans}})\,[\sqrt{P_{\text{MW}}}\,B_{\text{m}}\,Q\,n_{\text{B}}\,S(S+1)])$$ where the quantity notations possess the following meaning (whether it is instrumental or sample dependent it is presented in parentheses):

Quantity Symbol	Meaning/Short Desription

\(c\)	Point sample calibration factor (instrumental).

\(f(B_1,B_\text{m})\)	Spatial distribution of the microwave \(B_1\) and modulation amplitude within the cavity/probehead/resonator (instrumental).

\(G_{\text{R}}\)	Receiver gain (commonly in \(\text{dB}\) units (instrumental)).

\(t_{\text{C}}\)	Conversion time (commonly in \(\text{ms}\)) which is an analogy to integration time in other spectroscopies (instrumental).

\(N_{\text{Scans}}\)	Number of scans/accumulations during the experiment (instrumental).

\(P_{\text{MW}}\)	Microwave power (instrumental).

\(B_{\text{m}}\)	Modulation amplitude.

\(Q\)	Q-Value or Q-Factor characterizing the resonator/cavity/probehead sensitivity (unitless and instrumental).

\(n_{\text{B}}\)	Boltzmann factor for temperature dependence (instrumental-sample).

\(S\)	Total electronic spin quantum number (sample). Commonly, for radicals \(S = 1/2\).

Almost all summarized quantities are instrument-dependent. Most of them correspond to the essential parameters for the experiment and can be easily acquired from the .DSC/.dsc/.par file(s). The Boltzmann factor describes the population of spin states by \(\exp{(\Delta \varepsilon)\,/\,(k_{\text{B}}\,T)}\), where \(\Delta \varepsilon\) denotes the energy difference between the basic spin states, \(k_{\text{B}}\) is the Boltzmann constant (available from syms) and the \(T\) represents the temperature in \(\text{K}\). For temperatures \(\geq 4\,\text{K}\) and continuous wave experiments where the \(\Delta \varepsilon = h\,\nu_{\text{MW}}^{}\) is constant, this factor may be very well estimated by the following formula $$n_{\text{B}} = h\,\nu_{\text{MW}}^{}\,/\,(2\,k_{\text{B}}\,T)$$ The term \((G_{\text{R}}\,t_{\text{C}}\,N_{\text{Scans}})\) actually corresponds to normalization constant which is available from quantify_EPR_Norm_const. Besides the above-described parameters which can be easily estimated there are however characteristics that requires precise calibration and usually are provided by the spectrometer manufacturers. The spatial distribution of the microwave, \(B_1\), and modulation amplitude, \(B_\text{m}\), influences the intensity of the sample predominantly along the \(y\)-axis direction (i.e. when moving the sample tube up or down within the cavity). Such intensity distribution, depending on the cavity/probehead for different sample length and positions, can be approximated by a polynomial (see the fn.B1.Bm.fit.y argument) that is supplied by the manufacturer as well (the coefficients of the polynomial can be sometimes found in .DSC/.dsc/.par file(s)). For quantitative purposes, such polynomial is integrated over the length of the sample.

References

eatonQepr2010eprscope

RWeberXenon2011eprscope

hirschCapill2023eprscope

mazurAnalysis2000eprscope

portisElectro1953eprscope

mailerQuant1977eprscope

See also

Examples

if (FALSE) { # \dontrun{
## quantitative analysis (determining the
## radical concentration `c.M`) of a sample measured
## at different temperatures
## all data summarized in `data.tidy.integ`
data.quant <- mapply(function(x,y)
  {quantify_EPR_Abs(x,
    instrum.params = c(PmW = 2.518,
                       BmmT = 5.4e-02,
                       TK = NULL, ## 298 K
                       mwGHz = 9.530265),
    path_to_dsc_par = NULL,
    origin = NULL,
    tube.sample.id.mm = 2.86,
    fill.sample.h.mm = 23,
    Temp.K = y)$c.M
    },
  data.tidy.integ$Area,
  data.tidy.integ$Temperature_K
  )
} # }