Calculation of Spectral Linewidth ( e.g. \(\Delta B_{pp}\)) of the EPR/ENDOR Spectrum

Calculating the peak-to-peak (distance between of the x-axis projection of "min" and "max" derivative intensities) linewidth of an EPR/ENDOR spectrum.

Usage

eval_DeltaXpp_Spec(
  data.spectr,
  x = "B_mT",
  Intensity = "dIepr_over_dB",
  xlim = NULL
)

Arguments

data.spectr

EPR/ENDOR spectrum data frame object with magnetic flux density B (in mT or G) or g-Value or RF (in MHz) column/variable and that of the derivative dIepr_over_dB Intensity. Index column may be included as well.

x

Character string pointing to name of the x-axis/column/variable (in the original data.spectr) like magnetic flux density \(B\), \(g\)-Value or \(RF\) (radio frequency), default: x = "B_mT".

Intensity

Character string pointing to name of the intensity column/variable, in the original data.spectr, if other than dIepr_over_dB name/label is used (e.g. for simulated spectra). Default: Intesity = "dIepr_over_dB".

xlim

Numeric vector corresponding to lower and upper limit of the selected x region, such as xlim = c(3495.4,3595.4) (B in G) or xlim = c(12.5,21.2) (RF in MHz) or xlim = c(2.004,2.001) (g dimensionless). Default: xlim = NULL (corresponding to the entire x range).

Value

Numeric value difference of the x-axis quantity like B,g,RF (the absolute value) corresponding to minimum and maximum of the derivative intensity (dIepr_over_dB) in EPR/ENDOR spectrum.

See also

Examples

## loading the aminoxyl radical CW EPR spectrum:
aminoxyl.data.path <-
  load_data_example(file = "Aminoxyl_radical_a.txt")
aminoxyl.data <-
  readEPR_Exp_Specs(aminoxyl.data.path,
                    qValue = 2100)
#
## evaluation of the central linewidth (∆Bpp in `mT`):
eval_DeltaXpp_Spec(aminoxyl.data,
                   xlim = c(348.345,350.450))
#> [1] 0.541
#
## plot interactive spectrum
plot_EPR_Specs2D_interact(aminoxyl.data)

#
## the linewidth ∆Bpp (in `mT`) may be directly
## checked from the interactive spectrum above
## as a difference between the minimum and maximum
## of the `dIepr_over_dB` central line intensity:
349.648-349.107
#> [1] 0.541
#
## loading the perinaphthenyl (PNT) CW ENDOR spectrum:
pnt.endor.data.path <-
  load_data_example(file = "PNT_ENDOR_a.txt")
pnt.endor.data <-
  readEPR_Exp_Specs(pnt.endor.data.path,
                    col.names = c("index",
                                  "RF_MHz",
                                  "dIepr_over_dB"),
                   x.unit = "MHz")
#
## evaluation of the fourth linewidth (∆freq.(pp)
## in `MHz`) in the ENDOR spectrum:
eval_DeltaXpp_Spec(pnt.endor.data,
                  x = "RF_MHz",
                  xlim = c(22.38,24.54)
                  )
#> [1] 0.12
#
## plot interactive ENDOR spectrum:
plot_EPR_Specs2D_interact(pnt.endor.data,
                          x = "RF_MHz",
                          x.unit = "MHz")

#
## the linewidth (∆freq.(pp) in `MHz`) may be
## directly checked from the previous interactive
## CW ENDOR spectrum as a difference between
## the minimum and maximum of the `dIepr_over_dB`
## 4th line intensity:
23.42-23.30
#> [1] 0.12